Remove migration dependencies and update Backend name.
