+++ package.json
@@ -3,7 +3,6 @@
   "version": "3.1.0",
   "private": true,
   "dependencies": {
-    "@ant-design/compatible": "^1.0.8",
     "@ant-design/icons": "^4.2.2",
     "antd": "^4.6.5",
     "babel-plugin-import": "^1.13.0",
@@ -11,7 +10,7 @@
     "jquery": "^3.4.1",
     "less": "^3.11.0",
     "less-loader": "^7.0.1",
-    "moment": "^2.24.0",
+    "moment": "^2.29.0",
     "react": "^16.12.0",
     "react-app-rewired": "^2.1.5",
     "react-columns": "^1.2.1",
@@ -20,8 +19,7 @@
     "react-favicon": "0.0.18",
     "react-highlight-words": "^0.16.0",
     "react-router-dom": "^5.1.2",
-    "react-scripts": "^3.3.0",
-    "styled-components": "^5.0.1"
+    "react-scripts": "^3.3.0"
   },
   "scripts": {
     "start": "PORT=4200 react-app-rewired start",
